Move config to pipes-rs
