+++ crates/model/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod config;
 pub mod direction;
 pub mod pipe;
 pub mod position;
